Added level layout movement, way correction and deleted zoom
Level screen movement means that you can drag the screen to move the 
layout, to reach non-visible blocks.

Zoom was deleted because it was problematic, specifically when combined 
with screen movement, because it was making some unsightly movements, 
and solve them is not easy. In addition, I think that it's not necesary; 
screen movement fits better in the game